+++ lib/app/bloc/app_bloc.dart
@@ -1,0 +1,13 @@
+import 'package:bloc/bloc.dart';
+import 'package:equatable/equatable.dart';
+
+part 'app_event.dart';
+part 'app_state.dart';
+
+class AppBloc extends Bloc<AppEvent, AppState> {
+  AppBloc() : super(AppInitial()) {
+    on<AppEvent>((event, emit) {
+      // TODO: implement event handler
+    });
+  }
+}
