+++ lib/app/bloc/app_event.dart
@@ -1,0 +1,8 @@
+part of 'app_bloc.dart';
+
+abstract class AppEvent extends Equatable {
+  const AppEvent();
+
+  @override
+  List<Object> get props => [];
+}
