+++ lib/app/bloc/app_state.dart
@@ -1,0 +1,10 @@
+part of 'app_bloc.dart';
+
+abstract class AppState extends Equatable {
+  const AppState();
+  
+  @override
+  List<Object> get props => [];
+}
+
+class AppInitial extends AppState {}
